+++ unit_tests/base.py
@@ -134,20 +134,3 @@ class FakeLogger(object):
     
     def debug(self, msg):
         self.messages['debug'].append(msg)
-
-
-# This function was stolen from repoze.who.tests:
-def encode_multipart_formdata(fields):
-    BOUNDARY = '----------ThIs_Is_tHe_bouNdaRY_$'
-    CRLF = '\r\n'
-    L = []
-    for (key, value) in fields:
-        L.append('--' + BOUNDARY)
-        L.append('Content-Disposition: form-data; name="%s"' % key)
-        L.append('')
-        L.append(value)
-    L.append('--' + BOUNDARY + '--')
-    L.append('')
-    body = CRLF.join(L)
-    content_type = 'multipart/form-data; boundary=%s' % BOUNDARY
-    return content_type, body